[refactor] simplify serde code dependencies
