--- include/ex_actor/internal/actor_ref_serialization/actor_ref_deserialization_info.h
@@ -1,26 +1,0 @@
-// Copyright 2025 The ex_actor Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include "ex_actor/internal/actor.h"
-#include "ex_actor/internal/network.h"
-
-namespace ex_actor::internal {
-struct ActorRefDeserializationInfo {
-  uint32_t this_node_id = 0;
-  std::function<TypeErasedActor*(uint64_t)> actor_look_up_fn;
-  MessageBroker* message_broker = nullptr;
-};
-}  // namespace ex_actor::internal
